+++ test/libsolidity/syntaxTests/parsing/payable_without_arguments.sol
@@ -1,0 +1,7 @@
+contract C {
+	function f() public pure {
+		address payable q = payable;
+	}
+}
+// ----
+// ParserError: (70-71): Expected address to be converted to address payable.